PhotoService/requirements.txt
grpcio-tools
Minio